third_party/WebKit/LayoutTests/fast/css/invalidation/targeted-attribute-scope-expected.txt

Test invalidation set scope for attribute features.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 0
PASS successfullyParsed is true

TEST COMPLETE
